Create css styles for new webpage
Format webpage using flexbox, dividing the page into 6 main blocks.

Space header content using space-around justification, with logo on left and
other links to the right.

Add some random title and text to block two, and justify to center.
Create a placeholder box for an image to be added later.

Create a title for block three, and center the cards with their
respective text.

Create a quote and center the content, while aligning quote author to
the right.

Add a call to action box with a sign up button.

Finish with copyright and date, centered at bottom of the page.